refactor: decompose SessionManager into focused services (ARC-3)
Extract two focused service classes from the 1,747-line SessionManager:

- SessionTranscripts: JSONL reading, caching, pagination, summaries (313 lines)
  Owns parsedEntriesCache, getCachedEntries, readMessages, readMessagesForMonitor,
  readTranscript, readTranscriptCursor, getSummary

- SessionDiscovery: discovery polling, session map sync, filesystem scan (321 lines)
  Owns pollTimers, discoveryTimeouts, startDiscoveryPolling, stopDiscoveryPolling,
  syncSessionMap, maybeDiscoverFromFilesystem, cleanSessionMapForWindow,
  purgeStaleSessionMapEntries

SessionManager retains lifecycle (create/kill), persistence (load/save/encrypt),
terminal interaction (send/escape/interrupt), and health monitoring.
Delegates to extracted services via composition with dependency injection.

session.ts reduced from 1,747 to 1,321 lines (24% reduction).
Public API unchanged — all consumers see the same SessionManager interface.

Closes #1696
